refactor(chat): group message relates
