+++ src/main/java/sleppynavigators/studyupbackend/domain/chat/generator/message/ChallengeCompleteSystemMessageGenerator.java
@@ -1,4 +1,4 @@
-package sleppynavigators.studyupbackend.domain.chat.systemmessage;
+package sleppynavigators.studyupbackend.domain.chat.generator.message;
 
 import org.springframework.stereotype.Component;
 import sleppynavigators.studyupbackend.domain.event.ChallengeCompleteEvent;
