+++ src/main/java/sleppynavigators/studyupbackend/domain/chat/generator/message/SystemMessageGenerator.java
@@ -1,4 +1,4 @@
-package sleppynavigators.studyupbackend.domain.chat.systemmessage;
+package sleppynavigators.studyupbackend.domain.chat.generator.message;
 
 import sleppynavigators.studyupbackend.domain.event.EventType;
 import sleppynavigators.studyupbackend.domain.event.SystemMessageEvent;
